requirements.txt
requests
PyYAML
typer[all]
rich
